requirements.txt
streamlit
openai
PyPDF2
python-docx
pandas